+++ spec/support/client_spec_helpers.rb
@@ -4,7 +4,6 @@ module ClientSpecHelpers
     CopyTunerClient.configure(false) do |config|
       config.api_key = 'abc123'
       config.s3_host = 'copy-tuner.com'
-      config.html_escape = true
     end
   end
 end